Edit login validation

+++ back-end/app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -8,29 +8,48 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
+use App\Models\User;
 
 class AuthenticatedSessionController extends Controller
 {
-    /**
-     * Handle an incoming authentication request.
-     */
-    public function store(LoginRequest $request): JsonResponse
-    {
-       
-        $request->authenticate();
 
-        $user=$request->user();
+    public function store(LoginRequest $request): JsonResponse{
+
+
+        // Check if the user exists
+        $user = User::where('email', $request->email)->first();
+
+        if (!$user) {
+            return response()->json([
+                'error' => "This email doesn't exist."
+            ], 404);
+        }
+
+
+
+        // Check if the password is correct
+        if (!Auth::attempt($request->only('email', 'password'))) {
+            return response()->json([
+                'error' => "Wrong password."
+            ], 401); 
+        }
+
+
+
         $user->tokens()->delete();
-        $token=$user->createToken('api-token');
 
+        // Create a new token for the user
+        $token = $user->createToken('api-token');
+
+        // Return the response with user details and token
         return response()->json([
-            'user'=> $user,
-            'roles'=>$user->getRoleNames(),
-            "permissions" => $user->getAllPermissions(),
-            'token' => $token->plainTextToken
-
-       ] );
+            'user' => $user,
+            'roles' => $user->getRoleNames(),
+            'permissions' => $user->getAllPermissions(),
+            'token' => $token->plainTextToken,
+        ], 200);
     }
+
 
     /**
      * Destroy an authenticated session.
